Project_1_Kenyan_Market_ETL/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0
requests>=2.28.0
pytest>=7.0.0
faker>=18.0.0